// packages/maker/src/api/action/ActionSection.ts
/*
 * Copyright 2025 Palantir Technologies, Inc. All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { SectionConditionalOverride } from "./SectionConditionalOverride.js";

export type ActionSection = {
  id: string;
  displayName: string;
  parameters: Array<string>;
  defaultVisibility?: "visible" | "hidden";
  description?: string;
  columnCount?: 1 | 2;
  showTitleBar?: boolean;
  collapsedByDefault?: boolean;
  style?: "box" | "minimal";
  conditionalOverrides?: Array<SectionConditionalOverride>;
};